+++ src/Traits/BuyableImageSpatieV7.php
@@ -12,37 +12,9 @@
 namespace Vanilo\Support\Traits;
 
 /**
- * Adds Buyable Image support using Spatie's Laravel Media Library V7
+ * @deprecated use HasImagesFromMediaLibrary instead
  */
 trait BuyableImageSpatieV7
 {
-    /**
-     * Returns whether the item has an image
-     *
-     * @return bool
-     */
-    public function hasImage(): bool
-    {
-        return $this->getMedia()->isNotEmpty();
-    }
-
-    /**
-     * Returns the URL of the item's thumbnail image, or null if there's no image
-     *
-     * @return string|null
-     */
-    public function getThumbnailUrl(): ?string
-    {
-        return $this->getFirstMediaUrl('default', 'thumbnail');
-    }
-
-    /**
-     * Returns the URL of the item's (main) image, or null if there's no image
-     *
-     * @return string|null
-     */
-    public function getImageUrl(): ?string
-    {
-        return $this->getFirstMediaUrl();
-    }
+    use HasImagesFromMediaLibrary;
 }