+++ src/Traits/BuyableImageSpatieV8.php
@@ -11,23 +11,12 @@
 
 namespace Vanilo\Support\Traits;
 
+use Illuminate\Support\Collection;
+
 /**
- * Adds Buyable Image support using Spatie's Laravel Media Library V8
+ * @deprecated use HasImagesFromMediaLibrary instead
  */
 trait BuyableImageSpatieV8
 {
-    public function hasImage(): bool
-    {
-        return $this->getMedia()->isNotEmpty();
-    }
-
-    public function getThumbnailUrl(): ?string
-    {
-        return $this->getFirstMediaUrl('default', 'thumbnail');
-    }
-
-    public function getImageUrl(): ?string
-    {
-        return $this->getFirstMediaUrl();
-    }
+    use HasImagesFromMediaLibrary;
 }