+++ Changelog.md
@@ -6,6 +6,8 @@
 ##### 2020-XX-YY
 
 - Added PHP 8 support
+- Added `HasImagesFromMediaLibrary` trait for supporting the new `HasImages` interface
+- Deprecated the `BuyableImageSpatieV8` and `BuyableImageSpatieV7` traits
 
 ### 2.0.1
 ##### 2020-10-28